chop the test selector into a million tiny pieces
