+++ src/main/java/io/wispforest/owo/braid/widgets/object/entity/EntityDisplayMode.java
@@ -1,0 +1,5 @@
+package io.wispforest.owo.braid.widgets.object.entity;
+
+public enum EntityDisplayMode {
+    FIXED, VANILLA, CURSOR, VANILLA_CURSOR, NONE
+}
